@@ -1,8 +1,6 @@
 module sigs.k8s.io/security-profiles-operator
 
-go 1.23.0
-
-toolchain go1.23.3
+go 1.24
 
 require (
 	dario.cat/mergo v1.0.1
