@@ -17,7 +17,6 @@ limitations under the License.
 package seccompprofile
 
 import (
-	"context"
 	"errors"
 	"fmt"
 	"os"
@@ -115,7 +114,7 @@ func TestReconcile(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			t.Parallel()
 
-			gotResult, gotErr := tc.rec.Reconcile(context.Background(), tc.req)
+			gotResult, gotErr := tc.rec.Reconcile(t.Context(), tc.req)
 			if tc.wantErr != nil {
 				require.EqualError(t, gotErr, tc.wantErr.Error())
 			}
@@ -765,7 +764,7 @@ func TestResolveSyscallsForProfile(t *testing.T) {
 			sut.impl = mock
 
 			syscalls, err := sut.resolveSyscallsForProfile(
-				context.Background(), sp, sp.Spec.Syscalls, logr.Discard(), 0,
+				t.Context(), sp, sp.Spec.Syscalls, logr.Discard(), 0,
 			)
 			assert(syscalls, err)
 		})
